% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Bucket}
\alias{Bucket}
\title{Fusion Tables API Objects 
API for working with Fusion Tables data.}
\usage{
Bucket(color = NULL, icon = NULL, max = NULL, min = NULL,
  opacity = NULL, weight = NULL)
}
\arguments{
\item{color}{Color of line or the interior of a polygon in #RRGGBB format}

\item{icon}{Icon name used for a point}

\item{max}{Maximum value in the selected column for a row to be styled according to the bucket color, opacity, icon, or weight}

\item{min}{Minimum value in the selected column for a row to be styled according to the bucket color, opacity, icon, or weight}

\item{opacity}{Opacity of the color: 0}

\item{weight}{Width of a line (in pixels)}
}
\value{
Bucket object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:13:25
filename: /Users/mark/dev/R/autoGoogleAPI/googlefusiontablesv2.auto/R/fusiontables_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Bucket Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the minimum and maximum values, the color, opacity, icon and weight of a bucket within a StyleSetting.
}

