% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Column.baseColumn}
\alias{Column.baseColumn}
\title{Column.baseColumn Object}
\usage{
Column.baseColumn(columnId = NULL, tableIndex = NULL)
}
\arguments{
\item{columnId}{The id of the column in the base table from which this column is derived}

\item{tableIndex}{Offset to the entry in the list of base tables in the table definition}
}
\value{
Column.baseColumn object
}
\description{
Column.baseColumn Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Identifier of the base column. If present, this column is derived from the specified base column.
}
\seealso{
Other Column functions: \code{\link{Column}},
  \code{\link{column.insert}}, \code{\link{column.patch}},
  \code{\link{column.update}}
}

