% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Column}
\alias{Column}
\title{Column Object}
\usage{
Column(Column.baseColumn = NULL, baseColumn = NULL, columnId = NULL,
  columnJsonSchema = NULL, columnPropertiesJson = NULL,
  description = NULL, formatPattern = NULL, graphPredicate = NULL,
  name = NULL, type = NULL, validValues = NULL, validateData = NULL)
}
\arguments{
\item{Column.baseColumn}{The \link{Column.baseColumn} object or list of objects}

\item{baseColumn}{Identifier of the base column}

\item{columnId}{Identifier for the column}

\item{columnJsonSchema}{JSON schema for interpreting JSON in this column}

\item{columnPropertiesJson}{JSON object containing custom column properties}

\item{description}{Column description}

\item{formatPattern}{Format pattern}

\item{graphPredicate}{Column graph predicate}

\item{name}{Name of the column}

\item{type}{Type of the column}

\item{validValues}{List of valid values used to validate data and supply a drop-down list of values in the web application}

\item{validateData}{If true, data entered via the web application is validated}
}
\value{
Column object
}
\description{
Column Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Specifies the details of a column in a table.
}
\seealso{
Other Column functions: \code{\link{Column.baseColumn}},
  \code{\link{column.insert}}, \code{\link{column.patch}},
  \code{\link{column.update}}
}

