% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{column.update}
\alias{column.update}
\title{Updates the name or type of an existing column.}
\usage{
column.update(Column, tableId, columnId)
}
\arguments{
\item{Column}{The \link{Column} object to pass to this method}

\item{tableId}{Table for which the column is being updated}

\item{columnId}{Name or identifier for the column that is being updated}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}

Other Column functions: \code{\link{Column.baseColumn}},
  \code{\link{Column}}, \code{\link{column.insert}},
  \code{\link{column.patch}}
}

