% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Geometry}
\alias{Geometry}
\title{Geometry Object}
\usage{
Geometry(geometries = NULL, geometry = NULL, type = NULL)
}
\arguments{
\item{geometries}{The list of geometries in this geometry collection}

\item{geometry}{No description}

\item{type}{Type: A collection of geometries}
}
\value{
Geometry object
}
\description{
Geometry Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a Geometry object.
}

