% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{LineStyle}
\alias{LineStyle}
\title{LineStyle Object}
\usage{
LineStyle(strokeColor = NULL, strokeColorStyler = NULL,
  strokeOpacity = NULL, strokeWeight = NULL, strokeWeightStyler = NULL)
}
\arguments{
\item{strokeColor}{Color of the line in #RRGGBB format}

\item{strokeColorStyler}{Column-value, gradient or buckets styler that is used to determine the line color and opacity}

\item{strokeOpacity}{Opacity of the line : 0}

\item{strokeWeight}{Width of the line in pixels}

\item{strokeWeightStyler}{Column-value or bucket styler that is used to determine the width of the line}
}
\value{
LineStyle object
}
\description{
LineStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a LineStyle within a StyleSetting
}
