% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Point}
\alias{Point}
\title{Point Object}
\usage{
Point(coordinates = NULL, type = NULL)
}
\arguments{
\item{coordinates}{The coordinates that define the point}

\item{type}{Point: A point geometry}
}
\value{
Point object
}
\description{
Point Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a point object.
}

