% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Polygon}
\alias{Polygon}
\title{Polygon Object}
\usage{
Polygon(coordinates = NULL, type = NULL)
}
\arguments{
\item{coordinates}{The coordinates that define the polygon}

\item{type}{Type: A polygon geometry}
}
\value{
Polygon object
}
\description{
Polygon Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a polygon object.
}
