% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{query.sql}
\alias{query.sql}
\title{Executes a Fusion Tables SQL statement, which can be any of - SELECT- INSERT- UPDATE- DELETE- SHOW- DESCRIBE- CREATE statement.}
\usage{
query.sql(sql, hdrs = NULL, typed = NULL)
}
\arguments{
\item{sql}{A Fusion Tables SQL statement, which can be any of}

\item{hdrs}{Whether column names are included in the first row}

\item{typed}{Whether typed values are returned in the (JSON) response: numbers for numeric values and parsed geometries for KML values}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
\item https://www.googleapis.com/auth/fusiontables.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables, https://www.googleapis.com/auth/fusiontables.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}

