% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Sqlresponse}
\alias{Sqlresponse}
\title{Sqlresponse Object}
\usage{
Sqlresponse(columns = NULL, rows = NULL)
}
\arguments{
\item{columns}{Columns in the table}

\item{rows}{The rows in the table}
}
\value{
Sqlresponse object
}
\description{
Sqlresponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a response to a SQL statement.
}

