% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{style.insert}
\alias{style.insert}
\title{Adds a new style for the table.}
\usage{
style.insert(StyleSetting, tableId)
}
\arguments{
\item{StyleSetting}{The \link{StyleSetting} object to pass to this method}

\item{tableId}{Table for which a new style is being added}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}

Other StyleSetting functions: \code{\link{StyleSetting}},
  \code{\link{style.patch}}, \code{\link{style.update}}
}

