% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{StyleFunction.gradient.colors}
\alias{StyleFunction.gradient.colors}
\title{StyleFunction.gradient.colors Object}
\usage{
StyleFunction.gradient.colors()
}
\value{
StyleFunction.gradient.colors object
}
\description{
StyleFunction.gradient.colors Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}
\seealso{
Other StyleFunction functions: \code{\link{StyleFunction.gradient}},
  \code{\link{StyleFunction}}
}

