% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{StyleSetting}
\alias{StyleSetting}
\title{StyleSetting Object}
\usage{
StyleSetting(markerOptions = NULL, name = NULL, polygonOptions = NULL,
  polylineOptions = NULL, styleId = NULL, tableId = NULL)
}
\arguments{
\item{markerOptions}{Style definition for points in the table}

\item{name}{Optional name for the style setting}

\item{polygonOptions}{Style definition for polygons in the table}

\item{polylineOptions}{Style definition for lines in the table}

\item{styleId}{Identifier for the style setting (unique only within tables)}

\item{tableId}{Identifier for the table}
}
\value{
StyleSetting object
}
\description{
StyleSetting Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a complete StyleSettings object. The primary key is a combination of the tableId and a styleId.
}
\seealso{
Other StyleSetting functions: \code{\link{style.insert}},
  \code{\link{style.patch}}, \code{\link{style.update}}
}

