% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{StyleSettingList}
\alias{StyleSettingList}
\title{StyleSettingList Object}
\usage{
StyleSettingList(items = NULL, nextPageToken = NULL, totalItems = NULL)
}
\arguments{
\item{items}{All requested style settings}

\item{nextPageToken}{Token used to access the next page of this result}

\item{totalItems}{Total number of styles for the table}
}
\value{
StyleSettingList object
}
\description{
StyleSettingList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a list of styles for a given table.
}
