% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{table.importTable}
\alias{table.importTable}
\title{Imports a new table.}
\usage{
table.importTable(name, delimiter = NULL, encoding = NULL)
}
\arguments{
\item{name}{The name to be assigned to the new table}

\item{delimiter}{The delimiter used to separate cell values}

\item{encoding}{The encoding of the content}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}
