% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Table}
\alias{Table}
\title{Table Object}
\usage{
Table(attribution = NULL, attributionLink = NULL, baseTableIds = NULL,
  columnPropertiesJsonSchema = NULL, columns = NULL, description = NULL,
  isExportable = NULL, name = NULL, sql = NULL, tableId = NULL,
  tablePropertiesJson = NULL, tablePropertiesJsonSchema = NULL)
}
\arguments{
\item{attribution}{Attribution assigned to the table}

\item{attributionLink}{Optional link for attribution}

\item{baseTableIds}{Base table identifier if this table is a view or merged table}

\item{columnPropertiesJsonSchema}{Default JSON schema for validating all JSON column properties}

\item{columns}{Columns in the table}

\item{description}{Description assigned to the table}

\item{isExportable}{Variable for whether table is exportable}

\item{name}{Name assigned to a table}

\item{sql}{SQL that encodes the table definition for derived tables}

\item{tableId}{Encrypted unique alphanumeric identifier for the table}

\item{tablePropertiesJson}{JSON object containing custom table properties}

\item{tablePropertiesJsonSchema}{JSON schema for validating the JSON table properties}
}
\value{
Table object
}
\description{
Table Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a table.
}
\seealso{
Other Table functions: \code{\link{table.insert}},
  \code{\link{table.patch}}, \code{\link{table.update}}
}

