% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{table.replaceRows}
\alias{table.replaceRows}
\title{Replaces rows of an existing table. Current rows remain visible until all replacement rows are ready.}
\usage{
table.replaceRows(tableId, delimiter = NULL, encoding = NULL,
  endLine = NULL, isStrict = NULL, startLine = NULL)
}
\arguments{
\item{tableId}{Table whose rows will be replaced}

\item{delimiter}{The delimiter used to separate cell values}

\item{encoding}{The encoding of the content}

\item{endLine}{The index of the line up to which data will be imported}

\item{isStrict}{Whether the imported CSV must have the same number of column values for each row}

\item{startLine}{The index of the first line from which to start importing, inclusive}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}

