% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{TableList}
\alias{TableList}
\title{TableList Object}
\usage{
TableList(items = NULL, nextPageToken = NULL)
}
\arguments{
\item{items}{List of all requested tables}

\item{nextPageToken}{Token used to access the next page of this result}
}
\value{
TableList object
}
\description{
TableList Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a list of tables.
}

