% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{task.get}
\alias{task.get}
\title{Retrieves a specific task by its ID.}
\usage{
task.get(tableId, taskId)
}
\arguments{
\item{tableId}{Table to which the task belongs}

\item{taskId}{The identifier of the task to get}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
\item https://www.googleapis.com/auth/fusiontables.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables, https://www.googleapis.com/auth/fusiontables.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}

