% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_functions.R
\name{task.list}
\alias{task.list}
\title{Retrieves a list of tasks.}
\usage{
task.list(tableId, maxResults = NULL, pageToken = NULL, startIndex = NULL)
}
\arguments{
\item{tableId}{Table whose tasks are being listed}

\item{maxResults}{Maximum number of tasks to return}

\item{pageToken}{Continuation token specifying which result page to return}

\item{startIndex}{Index of the first result returned in the current page}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/fusiontables
\item https://www.googleapis.com/auth/fusiontables.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/fusiontables, https://www.googleapis.com/auth/fusiontables.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/fusiontables}{Google Documentation}
}

