% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusiontables_objects.R
\name{Template}
\alias{Template}
\title{Template Object}
\usage{
Template(automaticColumnNames = NULL, body = NULL, name = NULL,
  tableId = NULL, templateId = NULL)
}
\arguments{
\item{automaticColumnNames}{List of columns from which the template is to be automatically constructed}

\item{body}{Body of the template}

\item{name}{Optional name assigned to a template}

\item{tableId}{Identifier for the table for which the template is defined}

\item{templateId}{Identifier for the template, unique within the context of a particular table}
}
\value{
Template object
}
\description{
Template Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the contents of InfoWindow templates.
}
\seealso{
Other Template functions: \code{\link{template.insert}},
  \code{\link{template.patch}},
  \code{\link{template.update}}
}
